requirements.txt
web.py==0.37
pyzmq==13.0.0
